Further tests
Tested Ajax.request()'s parameter validating logic for when user passes
in the arguments one by one and fixed a bug where,when the url or method
parameters were of the incorrect type(not a string),the request would be
attempted anyway.
One final round of tests will soon be made.